the core init class
this is now without any pointers. there's one iterator, but no pointers,
and no other classes get to access the init's list.

it seems likely that the next version of ChoreKeeper will need to
access the process list, but I think it'll get a copy of the list,
not access to the list itself. the list is small, after all.