a class to parse json server specs and remember them
Launcher uses this to read the json, and passes the result on to Chore
and eventually Process; the class doesn't really do much itself

I ended up using property_tree for parsing json since the code is so
straightforward and allocation-free. the others I looked at provide
more, but don't look as good and why do we need more? if this ends up
being too limit, we can either extend this (e.g. with lexical_cast to
get more data types) or change to either json spirit or jsoncpp.

the error reporting is very, very sparse. I think that's okay for now.