teach the chorekeeper how to pick a victim when thrashing
the kernel won't do a proper job, so we have to it ourselves.  our
algorithm is: when the kernel thrashes for eight seconds without pause, we
try several strategies for picking something killable. our first (and
best) way is to look for the one which is furthest above its owner's
declared maximum RAM. our most novel one is to pick the process that's
suffering worst from the thrashing.