the code worked, but the test case broke randomly.
a directory is a set of files, not a list, so if we want to treat it as a
list (which the test case does, but real code not) we need to sort it.
this change does just that. unfortunately I'm not sure it works with older
boost such as we have on aew[abc]. we'll see in a moment.