new Process::launch(), more likely to launch reliably.
this shifts code back into c++. launching a service now involves three
chained processes: download, install and the process itself. ops may
configure how download and install happen, but the chaining is hardcoded.

Init::manage() now returns a pointer to the managed object, so that
it's possible to chain reliably. this is evil, evil, evil, but I'm
tired of being close to orgasm but not quite there. I want to use
nodee, now, and if a function with a non-void return type and an
imperative name is what it takes, then that's what we'll have.