add artifact checksumming again.
it's now possible to specify an md5 sum for the artifact. when nodee
downloads the artifact, or chooses not to, it will take the md5 sum
into account.

this is more or less pointless for production, but neat for development.
you can use the same version number again and again:

  {
    "url": "http://myworkstation.foo.bar/.../queue-1.2-yourname-shaded.jar",
    "artifact": "foo:queue:1.2-yourname",
    "md5": "f3baba589574fb9e2ffdea52edea6b1f",
    ...

Every time you launch, nodee will see that the MD5 sum is different from
the previous sum it saw, connect to your workstation and retrieve your
latest snapshot, and launch that.

Note that using version number 1.2-snapshot won't work, since other people
may also be working on queue-1.2 and launch their working versions using
the same cluster.