use proper locking instead of sleep(2)
the basic problem here was that wait() degenerates the wrong way (for our
purposes). instead of waiting for children to be born and then die, it
would return at once.

so, we have to wait with calling wait() until a child has been born, but
we ought to do it immediately once a child has been born, since many of
our children live only briefly.

this uses an (unnecessary) mutex and a condition variable to achieve that
end: if init can't wait because there are no children, it will wait for a
condition variable to be triggered, and we trigger than variable whenever
we add a child.