cleaner host reboot
the manic reboot was too painful. when something went wrong, regaining
control of the host was too much of a challenge.

this script allows five fast reboots, but after that, reboot slows down
enough to allow an admin to log in, and if that does happen, the reboot is
handed off to the admin.

much more pleasant.